Move current type by node tests to deprecated package
